test: 채팅 저장에 관련된 테스트를 ChatMessageStoreTest로 이동
